chainlit_lite_chatbot/pyproject.toml
[project]
name = "chainlit-lite-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.5.5",
    "google-auth>=2.40.2",
    "google-genai>=1.18.0",
    "jason>=0.1.7",
    "litellm>=1.71.3",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
